% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEEfun.R
\name{C1.new}
\alias{C1.new}
\title{Basic TPEE Calculation}
\usage{
C1.new(d, s, e, h1, h2)
}
\arguments{
\item{d}{dose/concentration value DoseX}

\item{s}{time since start tStartX}

\item{e}{time since end tEndX}

\item{h1}{half-life parameter for incline}

\item{h2}{half-life parameter for decline}
}
\description{
Reads in dose, start and end times, and two half-life values - one in, one out. Calculates concentration for inputs. Allows for two differing half-lives.
}
\examples{
C1.new()
}
