% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEEfun.R
\name{C1fun.2h}
\alias{C1fun.2h}
\title{TPEE Calculation across dataset subjects and regimens}
\usage{
C1fun.2h(thalf = NULL, dat)
}
\arguments{
\item{thalf}{Assumed pair of half-lives for the TPEE calculation. Default is NULL. Entering a single half-life results in a warning and OPEE calculation.}

\item{dat}{Dataset with three columns per exposure event/regimen following the naming conventions for X total regimens: Dose1--DoseX, tStart1--tStartX, tEnd1--tEndX}
}
\description{
Reads in dataset and input half-life. Uses lapply and C1fn.h to calculate concentration for each observation for each regimen given in the dataset. Sums each dose for each subject at each time point. Outputs vector of concentrations for all observations in the dataset (should match the event vector length).
}
\examples{
C1fun.2h()
}
