% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find1half.R
\name{find.1half}
\alias{find.1half}
\title{find.1half()}
\usage{
find.1half(h0 = NULL, k0 = NULL, datter, modtype = "cph", stdose = 1,
  covar = NULL, stratas = NULL, iterh = NULL, printer = NULL,
  txtadd = NULL)
}
\arguments{
\item{h0}{Initialized half-life parameter to use in the OPEE estimation process. Default is NULL, though an error will be returned if neither h0 nor k0 are provided.}

\item{k0}{Initial rate to use for effective exposure estimation process. Default is NULL. Preferable to input h0 instead, though algorithm uses k0.}

\item{datter}{Dataframe with the corresponding columns needed in the estimation algorithm process. Please see additional documentation for appropriate structure of the dataset, including the required column names and types.}

\item{modtype}{Whether to use survival or pooled logistic regression methods. Default is \code{modtype="cph"} indicating to use survival. If anything else, will assume "pool".}

\item{stdose}{Specify the default standard dose for which the effective exposure plateau is desired. Default assumes that this is a binary exposure with value of 1 implying exposure periods.}

\item{covar}{Whether or not to adjust for covariates during the half-life findind process. Input expects that these will match the variable names (columns of the dataframe object) that are to be use. Default is NULL.}

\item{stratas}{For survival outcomes that require separate strata, indicate the list of stratas here. Input expects that these will match the variable names (columns of the dataframe object). Default is NULL.}

\item{iterh}{Indicator for whether or not to track all iterations as the algorithm progresses}

\item{printer}{Indicator for whether or not to output (in the console) the iterations of the algorithm}

\item{txtadd}{Text to add onto print-out (if specified by \code{printer=1}). Used for simulations tracking during parallel processing.}
}
\value{
Multiple objects related to the algorithmic convergence and final result using a one-parameter effective exposure model. (More details to be added)
}
\description{
Analytic algorithm for finding the OPEE half-life and effect size.
}
