% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTruths.R
\name{fitTruth.cph}
\alias{fitTruth.cph}
\alias{fitTruth.pl}
\alias{pullTruth.cph}
\title{Parallel processing fits for true exposure parameters}
\usage{
fitTruth.cph(datin, a, b, c, up, dn, t1, t2, case)

fitTruth.pl(datin, a, b, c, up, dn, case, truep = 1)

pullTruth.cph(datin)
}
\arguments{
\item{datin}{Dataframe that consists of columns with the following specifications. Should be organized as one observation per subject per time, and should be in a survival dataset format for using fitTruth.cph().}

\item{a}{Pointer to the dataframe column with the true effective exposure for each individual at each time.}

\item{b}{Pointer to the dataframe column that provides the current dosing level. This is similar to the "currently" exposed model, without adjustment for past exposure, when the dose is binary.}

\item{c}{Pointer to the dataframe column that indicates "Ever" exposed}

\item{up}{Pointer to the dataframe column that indicates "Currently" exposed}

\item{dn}{Pointer to the dataframe column that indicates "Past" exposed}

\item{t1}{Pointer to the dataframe column with "time".}

\item{t2}{Pointer to the dataframe column with "tstop". Only relevant for the fitTruth.cph() function.}

\item{case}{Pointer to the dataframe column with "event"}

\item{truep}{Specifies the number of "lag" parameters to assign when calculating the AIC for the true models. Default is 1. *Look into this for the cox model fits*}
}
\value{
Results in a dataframe consisting of model results for the true parameter fits.
\describe{
   \item{Beta1.Est and Beta1.SE}{Estimated effect size or coefficient on the log-scale, with corresponding variability.}
   \item{Ratio, Ratio.95L, and Ratio.95U}{Resulting estimates on the Odds or Hazard Ratio scale.}
   \item{logL and AIC}{The corresponding model fit values}
   \item{modtype}{Whether pooled logistic regression (pool) or cox proportional hazards (cph) was used}
   \item{estalgo}{Specification for the true parameter fit}
   \item{Half.in, Half.out, h0, numiter, and RatioDose}{Additional columns provided in the output to be able to merge with algorithmic results. All should be "NA"}
}
}
\description{
Simulation-based model fitting for "true" exposure parameters.
}
\section{Functions}{
\itemize{
\item \code{fitTruth.cph}: Fits CoxPH models for the "truth" parameters.

\item \code{fitTruth.pl}: Fits Pooled logistic regression models for the "truth" parameters.

\item \code{pullTruth.cph}: Pulls the relevant columns for analyzing "truth" values from a survival dataset
}}

