% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HessCPH.R
\name{BetaH.SEcph}
\alias{BetaH.SEcph}
\alias{Beta2H.SEcph}
\title{Standard error approximations for Cox Proportional Hazard Model OPEE/TPEE results}
\usage{
BetaH.SEcph(datter, t1, t2, case, cv, s, h, stdose)

Beta2H.SEcph(datter, t1, t2, case, cv, s, hin, hout, stdose)
}
\arguments{
\item{datter}{dataset with columns for the survival outcomes (time, tstop, event), effective exposure components (Dose1 -- DoseX, tStart1 -- tStartX, tEnd1 -- tEndX), covariates and stratas}

\item{t1, t2}{column pointers for "time" and "tstop" in the \code{datter} dataframe}

\item{case}{column pointer for "event" in the \code{datter} dataframe}

\item{cv}{column pointers for all model covariates in the datter dataframe}

\item{s}{column pointers for all model strata in the datter dataframe. Only applicable for survival models}

\item{h, hin, hout}{Final derived estimated value for half-life (h) or incline/decline half-lives (hin, hout) parameters for which standard errors need to be approximated.}

\item{stdose}{Relative steady-state dose for the beta parameter interpretation}
}
\value{
The following values are computed and returned
\describe{
   \item{Beta1.HSE}{Hessian-derived standard error for the beta estimate. Tends to be slightly larger than the estimated standard error from Cox regression.}
   \item{Ratio.H95L and Ratio.H95U}{Lower and Upper bounds, respectively, for Hazard Ratio based on 95\% normal approximation using hessian-derived standard errors.}
   \item{In.HSE}{Hessian-derived standard error for incline half-life parameter estimate. When using the OPEE model inputs, this represents the single parameter half-life standard error.}
   \item{In.95L and In.95U}{Lower and Upper bounds, respectively, for incline half-life parameter based on 95\% normal approximation using hessian-derived standard errors. When using the OPEE model inputs, this represents the single parameter half-life 95\% normally-approximated bounds.}
   \item{Out.HSE}{Hessian-derived standard error for decline half-life parameter estimate. Returns NA for OPEE model estimation.}
   \item{Out.95L and Out.95U}{Lower and Upper bounds, respectively, for decline half-life parameter based on 95\% normal approximation using hessian-derived standard errors. Returns NA for OPEE model estimation.}
}
}
\description{
Calculates the Hessian-Derived standard errors for Beta, and Half-Life parameters (single half-life or incline/decline half-lives). Outputs a vector of values (see output descriptions).
}
\section{Functions}{
\itemize{
\item \code{BetaH.SEcph}: For calculating the Hessian-derived standard errors following the OPEE algorithm.

\item \code{Beta2H.SEcph}: For calculating the Hessian-derived standard errors following the TPEE algorithm.
}}

