% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HessPool.R
\name{BetaH.SEpl}
\alias{BetaH.SEpl}
\alias{Beta2H.SEpl}
\title{Standard error approximations for Pooled Logistic Regression Model OPEE/TPEE results}
\usage{
BetaH.SEpl(inh, datset, stdose = 1)

Beta2H.SEpl(inh, outh, datset, stdose = 1)
}
\arguments{
\item{inh, outh}{Final derived estimated value(s) for single half-life (inh) or incline/decline half-lives (inh,outh) parameters for which standard errors need to be approximated.}

\item{datset}{Dataset with columns for the outcome (event), effective exposure components (Dose1 -- DoseX, tStart1 -- tStartX, tEnd1 -- tEndX), and covariates.}

\item{stdose}{Relative steady-state dose for the beta parameter interpretation. Default is binary (1) exposure.}
}
\value{
The following values are computed and returned
\describe{
   \item{Beta1.HSE}{Hessian-derived standard error for the beta estimate. Tends to be slightly larger than the estimated standard error from Cox regression.}
   \item{Ratio.H95L and Ratio.H95U}{Lower and Upper bounds, respectively, for Hazard Ratio based on 95\% normal approximation using hessian-derived standard errors.}
   \item{In.HSE}{Hessian-derived standard error for incline half-life parameter estimate. When using the OPEE model inputs, this represents the single parameter half-life standard error.}
   \item{In.95L and In.95U}{Lower and Upper bounds, respectively, for incline half-life parameter based on 95\% normal approximation using hessian-derived standard errors. When using the OPEE model inputs, this represents the single parameter half-life 95\% normally-approximated bounds.}
   \item{Out.HSE}{Hessian-derived standard error for decline half-life parameter estimate. Returns NA for OPEE model estimation.}
   \item{Out.95L and Out.95U}{Lower and Upper bounds, respectively, for decline half-life parameter based on 95\% normal approximation using hessian-derived standard errors. Returns NA for OPEE model estimation.}
}
}
\description{
Calculates the Hessian-Derived standard errors for Beta, and Half-Life parameters (single half-life or incline/decline half-lives). Outputs a vector of values (see output descriptions). *Currently does not adjust for covariates*
}
\section{Functions}{
\itemize{
\item \code{BetaH.SEpl}: For calculating the Hessian-derived standard errors following the OPEE algorithm.

\item \code{Beta2H.SEpl}: For calculating the Hessian-derived standard errors following the TPEE algorithm.
}}

