% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDVecs.R
\name{makeDVecs}
\alias{makeDVecs}
\title{Simplified version of constructing dose regimen vectors.}
\usage{
makeDVecs(struct = 0, Cp.vec = c(1), ts.vec = c(0), tf.vec = c(900),
  intlen = 1, studyt = NULL)
}
\arguments{
\item{Cp.vec}{Vector of doses for the regimens. Default assumes the binary exposure plateau and that there is a single regimen of exposure. To have multiple exposures, include the same number of vector components in \code{Cp.vec}, \code{ts.vec}, and \code{tf.vec}.}

\item{ts.vec}{Vector of start times for the regimens. Default assumes the exposure was started at time 0.}

\item{tf.vec}{Vector of end times for the regimens. Default assumes the exposure continues through time=900. When \code{studyt} is not specified, the last specified end-time (last regimen) is used as the total study time.}

\item{intlen}{Increment time to use. Default is 1 time unit.}

\item{studyt}{Total study follow-up time. Default is \code{NULL} and will pull the last regimen stop time.}

\item{structure}{Structure indicator. If turned on (1), then additional regimen is added for time following discontinuation. Default is off (0), assuming that the total number of regimens is fixed and does not need additional follow-up.}
}
\value{
Outputs dataframe of study intervals with corresponding/useable dose regimen vectors, current dose vector.

\code{Dose1}...\code{DoseX} Columns indicating the overall doses for each regimen. Repeated throughout for computational ease.

\code{tStart1}...\code{tStartX} Columns indicating the time since starting the specific regimen -- depends on the point in the trajectory. i.e. Takes a value of 0 for times prior to initiation and increments parallel with time following initiation.

\code{tEnd1}...\code{tEndX}  Columns indicating the time since discontinuing the specific regimen -- depends on the point in the trajectory. i.e. Takes a value of 0 for times prior to start of regimen and while regimen is "on". Increments parallel to time following discontinuation.

\code{time} Column for the study time at the observation.

\code{currD} Column with value for the current regimen dose for the subject-time-specific observation.

\code{everD} Column indicating whether any exposure has occurred as of (prior to and including) the subject-time-specific observation.
}
\description{
Reads in dose, start and end times, and half-life. Calculates concentration for inputs.
}
\examples{
makeDVecs()
}
