% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickCPH.R
\name{quick.cph}
\alias{quick.cph}
\title{Fast Cox Proportional Hazard Model fitter for streamlined simulations and algorithms}
\usage{
quick.cph(C.in, dat, t1, t2, case, strats = NULL, covs = NULL,
  offs = NULL, inits = NULL, wts = NULL, rnames = NULL,
  cntrl = coxph.control(), type = 1)
}
\arguments{
\item{C.in}{Vector of values for the effective exposure that corresponds in length and order with the Surv(time,tstop,event) outcome matrix.}

\item{dat}{Dataframe or matrix with pointers (t1, t2, case, covs, strats) specifying the relevant columns to use in the model.}

\item{t1}{Pointer to the "time" column.}

\item{t2}{Pointer to the "tstop" column.}

\item{case}{Pointer to the "event" column.}

\item{strats}{Vector of pointer values for the stratifying columns. Default is none.}

\item{covs}{Vector of pointer values for the covariate columns. Default is none.}

\item{offs}{Input specification of "offset" for agreg.fit(). Default is NULL.}

\item{inits}{Input specification of "init" for agreg.fit(). Default is NULL.}

\item{wts}{Input specification of "weights" for agreg.fit(). Default is NULL.}

\item{rnames}{Input specification of "rownames" for agreg.fit(). Default is NULL.}

\item{cntrl}{Default control setup for coxph models (coxph.control()).}

\item{type}{Default (1) specifies using Breslow's handling of tied events.}
}
\value{
Outputs a model of class="cph" with the log-likelihood (element 1), the beta estimate for C.in (element 2) and the corresponding std. error (element 3). Elements in the object are not labeled for speed purposes.
\describe{
   \item{output[1]=logLik}{Model fit log-likelihood value.}
   \item{output[2]=beta.C}{Model fit beta estimate for effective exposure.}
   \item{output[3]=beta.C.se}{Model fit standard error (sqrt(var[1,1])) for the beta estimate.}
}
}
\description{
Fast fitter for coxph models using the agreg.fit() function and column pointers. Defaults to Breslow's handling of ties.
}
