% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solves.R
\name{solve.red}
\alias{solve.red}
\title{Time to reduction in excess hazard}
\usage{
solve.red(beta, s.val, e.val, half.in, half.out)
}
\description{
Calculates the time to \% reduction based on initial and end-point hazard ratios
}
\examples{

solve.red(beta=log(2.63),s.val=6.5447956,e.val=2.5582798,half.in=5.85,half.out=5.85)
}
