% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummFun.R
\name{getSummary}
\alias{getSummary}
\alias{returnOR}
\alias{returnHR}
\title{Functions for returning summary statistics from model objects}
\usage{
getSummary(mod, dose = 1)

returnOR(modspec, coefname)

returnHR(modspec, coefname)
}
\arguments{
\item{mod}{Model object. Can have the \code{class} "speedglm", "coxph", or "cph".}

\item{dose}{Steady-state plateau for the beta parameter interpretation}

\item{modspec}{Model object to use}

\item{coefname}{Exact name of the variables for which Odds or Hazard ratio should be calculated. May include a vector of multiple coefficients.}
}
\value{
Estimated coefficients (with standard errors), exponentiated values (Ratio and 95\% CI), and log-likelihood for the model.

returnHR() and returnOR() functions additionally output the AIC for the model fit
}
\description{
Functions for returning summary statistics from model objects
}
