% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEEfun.R
\name{C1fn.h}
\alias{C1fn.h}
\title{Basic OPEE Calculation}
\usage{
C1fn.h(d, s, e, h)
}
\arguments{
\item{d}{dose/concentration value DoseX}

\item{s}{time since start tStartX}

\item{e}{time since end tEndX}

\item{h}{half-life parameter}
}
\value{
Outputs a numeric vector of values for concentration that is equal in length to the number of rows/observations in the dataframe \code{dat}.
}
\description{
Reads in dose, start and end times, and half-life. Calculates concentration for inputs.
}
\examples{
C1fn.h()
}
