% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEEfun.R
\name{C1fn.h}
\alias{C1fn.h}
\alias{C1.new}
\title{Basic OPEE/TPEE Calculation}
\usage{
C1fn.h(d, s, e, h)

C1.new(d, s, e, h1, h2)
}
\arguments{
\item{d}{dose/concentration value DoseX}

\item{s}{time since start tStartX}

\item{e}{time since end tEndX}

\item{h}{half-life parameter}

\item{h1, h2}{half-life parameters for incline and decline, respectively}
}
\value{
Outputs a single value for effective exposure concentration.
}
\description{
Reads in dose, start and end times, and half-life (or half-lives). Calculates concentration for inputs.
}
\section{Functions}{
\itemize{
\item \code{C1fn.h}: Uses the one-parameter specification.

\item \code{C1.new}: Allows for two differing half-lives and calculates the effective exposure concentration based on incline and decline parameters.
}}

