% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSurv.R
\name{full.surv}
\alias{full.surv}
\title{Prepare dataset for survival analysis}
\usage{
full.surv(datin)
}
\arguments{
\item{datin}{A dataframe that must have "time", "tstop" and "event" columns with data points for all subjects at all event times.}
}
\value{
Outputs full dataset at event times
}
\description{
Prepare dataset for survival analysis
}
\seealso{
makeDat.cph
}
